+++ src/crypto/SHA256Tester.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 class SHA256Impl;
 
 class SHA256Tester
@@ -7,5 +9,5 @@ class SHA256Tester
 public:
 	SHA256Tester() = delete;
 	static bool verify(const SHA256Impl* impl);
-	static int benchmark(const SHA256Impl* impl);
+	static int64_t benchmark(const SHA256Impl* impl);
 };
